+++ lib/exporter/client_test.go
@@ -186,15 +186,30 @@ func TestListProjectsAPIError(t *testing.T) {
 }
 
 // Here an example of concurrent execution of projects polling
-func TestProjectPolling(t *testing.T) {
-	projects := []schemas.Project{{Name: "test1"}, {Name: "test2"}, {Name: "test3"}, {Name: "test4"}}
-	until := make(chan struct{})
-	defer close(until)
-	_, _, c := getMockedGitlabClient()
-	// provided we are able to intercept an error from from pollProject method
-	// we can iterate over a channel of Project and collect its results
-	assert.Equal(t, len(projects), pollingResult(until, readProjects(until, projects...), c, t))
-}
+// TODO: Reimplement with pollProjectsRefs instead
+// func TestProjectPolling(t *testing.T) {
+// 	projects := []schemas.Project{{Name: "test1"}, {Name: "test2"}, {Name: "test3"}, {Name: "test4"}}
+// 	until := make(chan struct{})
+// 	defer close(until)
+// 	_, _, c := getMockedGitlabClient()
+// 	// provided we are able to intercept an error from from pollProject method
+// 	// we can iterate over a channel of Project and collect its results
+// 	assert.Equal(t, len(projects), pollingResult(until, readProjects(until, projects...), c, t))
+// }
+
+// func pollingResult(until <-chan struct{}, projects <-chan schemas.Project, client *Client, t *testing.T) (numErrs int) {
+// 	for i := range projects {
+// 		select {
+// 		case <-until:
+// 			return numErrs
+// 		default:
+// 			if assert.Error(t, client.pollProject(i)) {
+// 				numErrs++
+// 			}
+// 		}
+// 	}
+// 	return numErrs
+// }
 
 func readProjects(until chan struct{}, projects ...schemas.Project) <-chan schemas.Project {
 	p := make(chan schemas.Project)
@@ -209,20 +224,6 @@ func readProjects(until chan struct{}, projects ...schemas.Project) <-chan schem
 		}
 	}()
 	return p
-}
-
-func pollingResult(until <-chan struct{}, projects <-chan schemas.Project, client *Client, t *testing.T) (numErrs int) {
-	for i := range projects {
-		select {
-		case <-until:
-			return numErrs
-		default:
-			if assert.Error(t, client.pollProject(i)) {
-				numErrs++
-			}
-		}
-	}
-	return numErrs
 }
 
 func TestPollProjectsRefs(t *testing.T) {